Add files via upload
This is a website of an online grocery shopping platform developed using HTML, CSS, and JS. This project involves only front-end coding languages. This was a group project and we will try to make it better. Any feedback is welcome on the project. One last thing, I am uploading the images used in this project but we have added the images using the URLs so there should be no need to add them again, for reference, we will be uploading the images.